Security Advisory - linux - CVE-2011-3191 - cifs: fix possible memory corruption in CIFSFindNext
CQID: WIND00353079

1/1 [
Author: Ming Liu
Email: ming.liu@windriver.com
Subject: cifs: fix possible memory corruption in CIFSFindNext
Date: Mon, 2 Jul 2012 09:53:57 +0800

commit 9438fabb73eb48055b58b89fc51e0bc4db22fabd upstream.

The name_len variable in CIFSFindNext is a signed int that gets set to
the resume_name_len in the cifs_search_info. The resume_name_len however
is unsigned and for some infolevels is populated directly from a 32 bit
value sent by the server.

If the server sends a very large value for this, then that value could
look negative when converted to a signed int. That would make that
value pass the PATH_MAX check later in CIFSFindNext. The name_len would
then be used as a length value for a memcpy. It would then be treated
as unsigned again, and the memcpy scribbles over a ton of memory.

Fix this by making the name_len an unsigned value in CIFSFindNext.

Cc: <stable@kernel.org>
Reported-by: Darren Lavender <dcl@hppine99.gbr.hp.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Integrated-by: Ming Liu <ming.liu@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>