fsl_p50x0: Add a work-around for erratum A-004510 in the Freescale PAMU device driver
CQID: WIND00355634

1/1 [
Author: Zhang Xiao
Email: xiao.zhang@windriver.com
Subject: Add a work-around for erratum A-004510 in the Freescale PAMU device driver
Date: Tue, 10 Jul 2012 11:24:30 +0800

Provide libedit_20110709-3.0.bb to be used for 5.0 build system.

Erratum A-004510 says that under certain load conditions, modified cache lines
can be discarded, causing data corruption.

The work-around for erratum A-004510 says that only one core at a time can
generate transactions that snoop the PAMU, and use a spinlock that protects
reads/writes to the PAACT ensures this.

Integrated-by: Tiejun Chen <tiejun.chen@windriver.com>
Integrated-by: Zhang Xiao <xiao.zhang@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>