Unconditionally enable oprofile backtrace for powerpc architectures
RALLY: US18075

1/1 [
Author: Dave Lerner
Email: dave.lerner@windriver.com
Subject: Unconditionally enable oprofile backtrace for powerpc architectures
Date: Wed, 18 Jul 2012 13:24:45 -0400

Make the assignment of the backtrace callback pointer unconditional.
The assignment was done after checking that the core variant and
PMC information was correctly setup, but backtrace logic for oprofile
doesn't rely on performance monitor registers.

In the original logic, oprofile is not disabled when PMC hardware
is not available.  Instead for the qemuppc target, oprofile sampling
triggers with the jiffy timer interrupt.  Since the original logic
enabled oprofile on a target without PMC hardware, there is no reason
not to enable backtrace when the oprofile sample is taken, the logic
added by this patch.

Signed-off-by: Dave Lerner <dave.lerner@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>