routerstationpro: import new patches from openwrt v3.3 kernel
Import of the openwrt files from trunk, as of this commit:

 ----------
commit c1d79f64eed0a7ac36b5b9bca52275b397bec424
Author: nbd <nbd@3c298f89-4303-0410-b956-a3cf2f4a3e73>
Date:   Mon Jul 16 16:26:51 2012 +0000

    uboot-ar71xx: fix compile on recent mac os x versions

    git-svn-id: svn://svn.openwrt.org/openwrt/trunk@32750 3c298f89-4303-0410-b956-a3cf2f4a3e73
 ----------

Due to the nature of how the openwrt source is managed, new
files are found separate from the commits which actually
make use of them, so that unused stuff could not easily
be left behind, but it is what it is.

Path to files in the repo is:
	target/linux/generic/files
	target/linux/ar71xx/files

Repo is: git://nbd.name/openwrt.git

The process used was creating a booting kernel based on a
v3.3 baseline and a subset of the openwrt patches (there
are over 200), with the obvious (aka "easy") cull made.
This subset was then forward ported to v3.4 -- it is by
no means the absolute minimum it could be; things like
ar93xx/95xx support could also be culled, or some of the
board specific init files, but given this board is already
EOL, I can't justify spending additional time doing that.

Some of the patches (ones with Patchworks links and SOB via
Ralf) made it upstream in v3.5-rc1, and I've added the
upstream IDs to those where appropriate.  However they were
not compared one-by-one to verify what (if any) deltas were
introduced in the backport to openwrt's 3.3 kernel.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>