routerstationpro: specify chosen filesystems in audit input.
Call out the two filesystems used by this specific BSP in
the hardware list.  While filesystems aren't generally
hardware items, boards with MTD devices will need certain
specific filesystems, so there is a hardware linkage.

Regardless, the point of the audit is to inform people in
a way that lets them think about their choices.  We've
done that here, and know we need these.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>