kconfig: add tty/serial file to hardware listing
A scan of this file reveals that it is full of UART specific
settings for various embedded platforms, and should be listed
in the hardware category.  This will also stop the audit from
warning that SERIAL_CORE and SERIAL_CORE_CONSOLE are being
set by BSPs as if it were non-hardware related.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>