kconfig: make CONFIG_BLOCK=y part of base config
This has largely no effect, since the default is y, and since
it is as fundamental as CONFIG_NET, it should be in the base.
Move it there and delete any other instances of it.  This will
resolve any BSP configuration audit warnings on CONFIG_BLOCK
caused by a BSP that includes scsi.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>