security/keys: fix bad patch selection in CVE backport
NOCQID

There was a patch used to convert to kfree_rcu, so that a follow
on patch could be used as-is for a cve fix.  However, there is
no kfree_rcu in the context of 2.6.34 (it was added in 3.0 kernels)
so this way of integrating the cve fix does not even compile.

Drop the conversion patch, and update the actual CVE patch to
use the code context of the 2.6.34 patch.

Cc: ming.liu@windriver.com
Cc: Yue.Tao@windriver.com
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>