fsl-e500mc: restore this bsp
NOCQID

This is a partial revert of commit:
07c9890e474e31bff28ccfd52c8d330fa12480da

"Sorry for the confusing, In Robin, the fsl_e500mc still uses the SDK
1.1 base. The SDK 1.2 is used to do Robin async.

"We though we are discussing Robin async, the Robin is almost GA, we will
not migrate fsl_e500mc to SDK 1.2.

"So we need to restore the SDK 1.1 base for Robin sync.

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>