meta: cleanup invalid/obselete 3.4 CONFIG options
The kernel configuration audit subsystem was previously focussed on
BSP/hardware specific options, since that is the part of configuration
that the end developer maintains.

But not auditing the non-hardware specific options on each run meant
that some options that do not exist in the kernel have remained in the
configuration fragments.

Removing them clarifies the fragments, and updating the audit to report
on non-hardware options ensures that they will stay clean.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>