intel-xeon-core: meta cleanup.
CQID: WIND00344708

It is the moment to clean up the -extra stuff from meta. Now that kexec,
kdump, ipmi, edac, ras and even kvm are prepared to be templatized, it is
not required to maintain them in bsp meta. Delete the duplicated and move
the necessary back to intel-xeon-core.scc and intel-xeon-core.cfg.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>