net/dpa: protect qman_irqsource_add/remove with preempt disabled
CQID: WIND00386256

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: net/dpa: protect qman_irqsource_add/remove with preempt disabled
Date: Wed, 31 Oct 2012 14:02:56 +0800

The qman_irqsource_add/remove are used to enable/disable a specific
interrupt source to assert the interrupt signal to the portal
which is bind to the current cpu. So we must make sure
that the pair of qman_irqsource_add/remove are run on the same
cpu.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>