lsi-acp34xx: support kexec/kdump
CQID: WIND00344712

1/5 [
Author: Wei Yang
Email: Wei.Yang@windriver.com
Subject: powerpc/47x: Kernel support for KEXEC
Date: Fri, 2 Nov 2012 17:04:31 +0800

commit 68343020031585f861d93e2f25589598feadaff3 upstream

This patch adds support for creating 1:1 mapping for the PPC_47x during
a KEXEC. The implementation is similar to that of the PPC440x which is
described here :

	http://patchwork.ozlabs.org/patch/104323/

PPC_47x MMU :

The 47x uses Unified TLB 1024 entries, with 4-way associative mapping
(4 x 256 entries). The index to be used is calculated by the MMU by
hashing the PID, EPN and TS. The software can choose to specify the way
by setting bit 0(enable way select) and the way in bits 1-2 in the TLB
Word 0.

Implementation:

The patch erases all the UTLB entries which includes the tlb covering
the mapping for our code. The shadow TLB caches the mapping for the
running code which helps us to continue the execution until we do
isync/rfi. We then create a tmp mapping for the current code in the
other address space (TS) and switch to it.

Then we create a 1:1 mapping(EPN=RPN) for 0-2GiB in the original
address space and switch to the new mapping.

TODO: Add SMP support.

Signed-off-by: Suzuki K. Poulose <suzuki@in.ibm.com>
Signed-off-by: Josh Boyer <jwboyer@gmail.com>
Integrated-by: Wei Yang <Wei.Yang@windriver.com>
]

2/5 [
Author: Suzuki Poulose
Email: suzuki@in.ibm.com
Subject: powerpc/47x: Enable CRASH_DUMP
Date: Sun, 15 Apr 2012 22:27:35 +0000

commit 47da421981571c69ef29740cc55fa7248682e167 upstream

Now that we have KEXEC and relocatable kernel working on 47x (!SMP)
enable CRASH_DUMP.

Signed-off-by: Suzuki K. Poulose <suzuki@in.ibm.com>
Signed-off-by: Josh Boyer <jwboyer@gmail.com>
Integrated-by: Wei Yang <Wei.Yang@windriver.com>
]

3/5 [
Author: Wei Yang
Email: Wei.Yang@windriver.com
Subject: powerpc/44x: Fix/Initialize PID to kernel PID before the TLB search
Date: Tue, 6 Nov 2012 14:10:17 +0800

commit f13bfcc6961a5c9f511c401292db522edcd0b061 upstream

Initialize the PID register with kernel pid (0) before we start
setting the TLB mapping for KEXEC. Also set the MMUCR[TID] to kernel
PID.

This was spotted while testing the kexec on ISS for 47x. ISS  doesn't
return a successful tlbsx for a kernel address with PID set to a user PID.
Though the hardware/qemu/simics work fine.

This patch is harmless and initializes the PID to 0 (kernel PID) which
is usually the case during a normal kernel boot. This would fix the kexec
on ISS for 440. I have tested this patch on sequoia board.

Signed-off-by: Suzuki K Poulose <suzuki@in.ibm.com>
Cc: Josh Boyer <jwboyer@gmail.com>
Signed-off-by: Josh Boyer <jwboyer@gmail.com>
Integrated-by: Wei Yang <Wei.Yang@windriver.com>
]

4/5 [
Author: Wei Yang
Email: Wei.Yang@windriver.com
Subject: powerpc/prom: remove the illegal reversed memory region
Date: Tue, 20 Nov 2012 14:40:40 +0800

When the MEMORY_START is larger than 0 we need to ensure that kernel
does not reserve the range from 0 to MEMORY_START in memblock. Because
the range is out of the available memory.

Signed-off-by: Wei Yang <Wei.Yang@windriver.com>
]

5/5 [
Author: Wei Yang
Email: Wei.Yang@windriver.com
Subject: powerpc/44x: kexec for SMP 47x
Date: Tue, 20 Nov 2012 18:20:11 +0800

Add SMP support for kexec on acp3400 board.
The implementation is similar to that of the 85xx which is described
here:
commit id: 933a41e419a954ef90605224e02c3ded78f3372 upstream
[
powerpc/85xx: kexec for SMP 85xx BookE systems

Adds support for kexec on 85xx machines for the BookE platform.
Including support for SMP machines
]

Signed-off-by: Wei Yang <Wei.Yang@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>