drop pch_uart-Correct-spin_lock-and-irqsave-usage-in-pch_.patch
NOCQID

Two things happened.  In the following commit I forgot to drop the
patch file when the patch directive was removed from
fri2-preempt-rt.scc:
4f34f01 squash serial-pch_uart-no-need-to-disable-irq-for-rt-kernel.patch with fri2 version

Then, when I merged yocto-master-edcb1523 I let the patch directive back
in because I was confused by the presence of the patch.  So Here I
drop the patch line from fri2-preempt-rt.scc, because this patch is on
rt common.

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>