sctp: ABORT if receive, reassmbly, or reodering queue is not empty while closing socket
CQID: WIND00395173

1/1 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: sctp: ABORT if receive, reassmbly, or reodering queue is not empty while closing socket
Date: Mon, 24 Dec 2012 10:12:59 +0800

commit cd4fcc704f30f2064ab30b5300d44d431e46db50 upstream

Trigger user ABORT if application closes a socket which has data
queued on the socket receive queue or chunks waiting on the
reassembly or ordering queue as this would imply data being lost
which defeats the point of a graceful shutdown.

This behavior is already practiced in TCP.

We do not check the input queue because that would mean to parse
all chunks on it to look for unacknowledged data which seems too
much of an effort. Control chunks or duplicated chunks may also
be in the input queue and should not be stopping a graceful
shutdown.

Signed-off-by: Thomas Graf <tgraf@infradead.org>
Acked-by: Vlad Yasevich <vladislav.yasevich@hp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>