xilinx-zynq: Add patches to support kexec/kdump
CQID: WIND00386039

1/3 [
Author: John Linn
Email: john.linn@xilinx.com
Subject: Xilinx: ARM: BSP: Update of UART for compressed image
Date: Fri, 23 Sep 2011 11:59:18 -0600

Taken from git tree git://git.xilinx.com/linux-xlnx.git:
commit 61f024f359311803c9b87aa4594abbba832366df

This change is for mach-zynq as this was not a problem in
mach-xilinx.

In QEMU, u-boot is not used and the kernel has to initialize
the UART to get any kernel output. The kernel assumes a boot
loader initializes the UART which won't work in QEMU or when
not using u-boot with a Xilinx JTAG probe.

Signed-off-by: Liming Wang <liming.wang@windriver.com>
]

2/3 [
Author: John Linn
Email: john.linn@xilinx.com
Subject: Xilinx: ARM: BSP: update memory initialization
Date: Wed, 18 Jul 2012 14:53:58 -0700

Taken from git tree git://git.xilinx.com/linux-xlnx.git:
commit 2fd319205a36892149ebad0e9d39c36f1f525ab2

This removes the reserving of memory for video and
also adds conditional compilation for the case where
the memory is not at 0 so as not to hang the kernel.

Signed-off-by: John Linn <john.linn@xilinx.com>
Signed-off-by: Liming Wang <liming.wang@windriver.com>
]

3/3 [
Author: Liming Wang
Email: liming.wang@windriver.com
Subject: arm/zynq: Reset hardware when initializing to avoid backtrace
Date: Fri, 28 Dec 2012 15:27:58 +0800

When starting kdump crash kernel, some pending interrupts of hardware
need to be clear. Otherwise pending interrupts may trigger the ethernet
driver backtrace because the irq system of kernel has not been ready then.

Signed-off-by: Liming Wang <liming.wang@windriver.com>
]

Signed-off-by: Liming Wang <liming.wang@windriver.com>
Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>