crypto: make cryptosoft SMP safe
CQID: WIND00318875

1/1 [
Author: hwang7
Email: huanhuan.Wang@windriver.com
Subject: crypto: make cryptosoft SMP safe
Date: Wed, 9 Jan 2013 11:35:47 +0800

When running test scripts which includes the following command
on fsl_p4080:

$ openssl speed -evp des -engine cryptodev -elapsed
$ openssl speed -evp des3 -engine cryptodev -elapsed
$ openssl speed -evp aes128 -engine cryptodev -elapsed

$ openssl speed -evp des -engine cryptodev -elapsed -multi 10
$ openssl speed -evp des3 -engine cryptodev -elapsed -multi 10
$ openssl speed -evp aes128 -engine cryptodev -elapsed -multi 10

The system will give out call trace or hang.

The root causes of this issue are:
1. The session list in fcrypt is not protected properly.
2. Though the swcr_sessions is protect by spin lock, but the three
types of session operation "create session, use session, delete
session" are not protected properly.

So:
1.add spin lock to protect session list in fcrypt
2.add state/ref fields to session, use session state to prevent from
deleting a session when it's in creating or using state, and use spin
lock to protect the swcr_sessions and session state.

Besides:
crypto_req_lock is deleted for the code it try to protect is dealing
with a dynamic allocated resource, and there is no SMP problems with
these codes. Futhermore, there is bug with the use of crypto_req_lock:
when this spin lock is held, there is kmalloc with GFP_KERNEL flag
set.

Signed-off-by: hwang7 <huanhuan.Wang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>