meta: drm-emgd.cfg: add configs for dependent routines
emgd driver utilizes common frame buffer routines to implement some
of the tasks. Making sure these routines are turned on in the kernel
config when emgd is selected.

This commit avoids these kind of kernel build errors.

| ERROR: "cfb_fillrect" [drivers/gpu/drm/emgd/emgd.ko] undefined!
| ERROR: "cfb_imageblit" [drivers/gpu/drm/emgd/emgd.ko] undefined!
| ERROR: "cfb_copyarea" [drivers/gpu/drm/emgd/emgd.ko] undefined!

And the I2C config bits are added to better support detection of modes
of monitors.

Signed-off-by: Nitin A Kamble <nitin.a.kamble@intel.com>