MIPS: XLP: handle irq on the starting cpu for kexecd && nosmp kernel
CQID: WIND00377560

1/1 [
Author: Yong Zhang
Email: yong.zhang@windriver.com
Subject: MIPS: XLP: handle irq on the starting cpu for kexecd && nosmp kernel
Date: Tue, 27 Nov 2012 15:28:29 +0800

By default, pic will be initialised to dispatch irqs to local cpus if
there is no matching ite; but this doesn't work for kernel which
is a kexecd one and have 'nosmp' set because we don't know on which
cpu a panic will happen.
So in such special condition we setup ite-0 to be the starting cpu
and let xlp_closest_match_cpumask() go back to ite-0 if no suitable
ite is found.

Signed-off-by: Yong Zhang <yong.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>