tipc: prevent dropped connections due to rcvbuf overflow
CQID: WIND00352520

1/1 [
Author: Erik Hugne
Email: erik.hugne@ericsson.com
Subject: tipc: prevent dropped connections due to rcvbuf overflow
Date: Thu, 4 Oct 2012 05:00:43 +0000

commit e57edf6b6dba975eceede20b4b13699d4e88cd78 upstream

When large buffers are sent over connected TIPC sockets, it
is likely that the sk_backlog will be filled up on the
receiver side, but the TIPC flow control mechanism is happily
unaware of this since that is based on message count.

The sender will receive a TIPC_ERR_OVERLOAD message when this occurs
and drop it's side of the connection, leaving it stale on
the receiver end.

By increasing the sk_rcvbuf to a 'worst case' value, we avoid the
overload caused by a full backlog queue and the flow control
will work properly.

This worst case value is the max TIPC message size times
the flow control window, multiplied by two because a sender
will transmit up to double the window size before a port is marked
congested.
We multiply this by 2 to account for the sk_buff and other overheads.

Signed-off-by: Erik Hugne <erik.hugne@ericsson.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>