sctp: fix to check the source address of COOKIE-ECHO chunk
CQID: WIND00397586

1/1 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: sctp: fix to check the source address of COOKIE-ECHO chunk
Date: Wed, 9 Jan 2013 10:39:59 +0800

commit de6becdc0844ff92b38ffd9f0c4db1d3de02835f upstream

SCTP does not check whether the source address of COOKIE-ECHO
chunk is the original address of INIT chunk or part of the any
address parameters saved in COOKIE in CLOSED state. So even if
the COOKIE-ECHO chunk is from any address but with correct COOKIE,
the COOKIE-ECHO chunk still be accepted. If the COOKIE is not from
a valid address, the assoc should not be established.

Signed-off-by: Wei Yongjun <yjwei@cn.fujitsu.com>
Signed-off-by: Vlad Yasevich <vladislav.yasevich@hp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>