intel-iommu: Extend PFN to 64 bits when DMAR >= 40 bits
CQID: WIND00399506

1/1 [
Author: Yang Shi
Email: yang.shi@windriver.com
Subject: intel-iommu: Extend PFN to 64 bits when DMAR >= 40 bits
Date: Wed, 30 Jan 2013 15:52:22 -0800

On x86_32, when the DMAR width is >= 40 bits and IOMMU is enabled,
unbinding a device from the host PCI driver will hang the associated
userspace process on a wait notifier that is never triggered.

This hang is caused on x86_32 when the DMAR width is >= 40 bits, since it
also means that the PFN is more than 32 bits. This cases the kernel to
infinitely loop in dma_pte_free_pagetable as follows:

  while (tmp && tmp + level_size(level) - 1 <= last_pfn)

Since the condition is always true.

Promote start_pfn and last_pfn parameter to 64 bit for dma_pte_clear_range
and dma_pte_free_pagetable. And, pass __DOMAIN_MAX_PFN to them as last_pfn.

Signed-off-by: Yang Shi <yang.shi@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>