oprofile: grsecurity changes results in x86-64 page fault
CQID: WIND00400436

1/1 [
Author: Amy Fong
Email: Amy.Fong@windriver.com
Subject: oprofile: grsecurity changes results in x86-64 page fault
Date: Mon, 11 Feb 2013 11:18:03 -0500

If profiling starts, shortly thereafter, we see an oops/page fault
in print_context_stack, or more specifically, at the line:

    while ( .... test to see if stack address is valid ...) {
         addr = *stack;
         stack++
    }

We appear to be walking out of range.

This patch modifies the start of the stack as follows:

    stack_start = (void *)((unsigned long)stack & ~(THREAD_SIZE-1));
to
    stack_start = (void *)current_thread_info();

The grsecurity, or more specifically the pax patches do the following:

"in particular we moved the thread info structure off the kernel stack
and into task struct (it's a long story, it was basically in response
to a particular kernel exploitation technique presented at HES'11 and
INFILTRATE'11 by Jon Oberheide and Dan Rosenberg, [1]).

<snip>

... depending on the circumstances
you can calculate the base address of the stack as you did above or use
the task_stack_page macro (which is the portable way across archs in linux
but it needs a struct task pointer)."

Signed-off-by: Amy Fong <Amy.Fong@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>