intel-xeon-core: use the default CONFIG_PHYSICAL_START=0x1000000.
CQID: WIND00405041

Workbench debugger uses system map file as the hint to find the
symbol location with compiling time address. But a relocatable
kernel will be loaded to another address and a proper relocation
fixup will be applied to. Thus, an inconsistent between runtime
address and compiling address happens. It is a long time to
await the fix by workbench debugger team, so we fix this in bsp,
even though workbench debugger will rarely fail with this fix,
except for dump-capture kernel.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>