MIPS fix access_ok
CQID: WIND00403430

1/1 [
Author: Yong Zhang
Email: yong.zhang@windriver.com
Subject: MIPS: fix access_ok()
Date: Wed, 20 Feb 2013 13:57:13 +0800

Current access_ok() will fail even if the address range is
valid when it reaches to the end of TASK_SIZE.
For example: addr = 0xfffffffff0; size = 16;
the real address range it want to access is 0xfffffffff0~0xfffffffff;
but addr + size = 0x10000000000 which we will not and can't access.
In current realization of access_ok(), the high bit will be 1
thus access_ok() indicates the operation is not allowed.

The bug is found in old kerenl(before vdso is realized) in
following typical call trace:
sys_mount()
  copy_mount_options()
    exact_copy_from_user()
When the parameter 'from' for exact_copy_from_user() residents in
the last page of the task's virtual address, such as stack.
But it's still in current kernel.

Signed-off-by: Yong Zhang <yong.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>