fsl-p50xx/fsl-t4xxx: enable math emulation and fix a typo
CQID: WIND00404272

the 'fsqrt' instruction is not implemented by e5500/e6500 core. so toolchain
should not use the instruction in 603e multilib. but its not fixed by
toolchain. so we temporarily enable the CONFIG_MATH_EMULATION option to
fix the issue. Additionally, I found a typo inside the configuration
file of p50xx, and fix it as well.

Signed-off-by: Wei Yang <Wei.Yang@windriver.com>