mips: fix signal.h SA_RESTORER build breakage
1/1 [
Author: Bruce Ashfield
Email: bruce.ashfield@windriver.com
Subject: mips: fix signal.h SA_RESTORER build breakage
Date: Mon, 25 Mar 2013 12:28:06 -0700

-stable commit cdef9602 [signal: always clear sa_restorer on execve]
assumes the presence of sa_restorer in the signal_handler structure.
This is not the case for MIPS o32, although the architecture does
defined SA_RESTORER. To fix this, we add the structure to the mips
local copy of signal.h.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>