fsl-t4xxx:CPU_HOTPLUG test cause system hang
CQID: WIND00408627

1/3 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: powerpc/fsl_book3e: add the support to online/offline a secondary thread separately
Date: Thu, 21 Mar 2013 15:17:46 +0800

In the currently implementation we can only offline all the threads
in a core completely and then online them sequentially. We can't do
as the following:
  offline cpu3(thread1 on core1)
  online cpu3

We can fix this limitation easily by invoking fsl_enable_threads
on the thread0 of the same core.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

2/3 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: powerpc/fsl_book3e: set the correct physical id for the secondary thread in cpu hotplug
Date: Wed, 20 Mar 2013 18:42:07 +0800

For the first booting, we would tweak the PIR register to get a
sequential physical id for all the threads. But in cpu hotplug,
the core will be reset when all the threads in a core are disabled.
That means the PIR register may contains the correct cpu physical id
or the hardware initialized value. And we don't have a way to
distinguish these two cases. We also can't set the cpu physical id
to a general register as what the kexec/kdump do since this register
will also be initialized after the core reset.

So we choose to get the cpu physical id from the PACA struct. Before
that we need to know the cpu logical id first. Fortunately we can
get this based on the current thread cpu logical id and TIR register.
This method also works for the booting.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

3/3 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: powerpc/fsl_book3e: disable the thread when it is offline
Date: Mon, 25 Mar 2013 10:05:31 +0800

When a thread is offline, we should disable it instead of spin here.
This should improve the performance of other threads which is still
running.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>