fix two bugs in simics support for t4240qds
CQID: WIND00408867

1/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: fsl-t4xxx: add the missing frag type for simics cfg
Date: Tue, 19 Mar 2013 14:12:43 +0800

The frag type is missed for the kconf command in the simics scc file
and then cause the following errors:
 basename: missing operand
 Try 'basename --help' for more information.
 [ERROR] kernel configuration fragment fragment '' cannot be found
 ERROR. A meta series could not be created for branch standard/base
 ERROR. Could not locate meta series for fsl-t4xxx
 ERROR. Could not apply patches for fsl-t4xxx.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

2/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: fsl-t4xxx: disalbe cgroups for simics
Date: Tue, 19 Mar 2013 14:37:28 +0800

The simics is not compatible with cgroups feature. So just disable
it when we build a kernel for simics.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>