inotify: fix double free/corruption of stuct user -- Security Advisory - linux - CVE-2011-1479
CQID: WIND00359329

1/1 [
Author: yanjun.zhu
Email: yanjun.zhu@windriver.com
Subject: inotify: fix double free/corruption of stuct user
Date: Wed, 13 Mar 2013 14:29:21 +0800

commit d0de4dc584ec6aa3b26fffea320a8457827768fc upstream

On an error path in inotify_init1 a normal user can trigger a double
free of struct user. This is a regression introduced by a2ae4cc9a16e
("inotify: stop kernel memory leak on file creation failure"). We fix
this by making sure that if a group exists the user reference is dropped
when the group is cleaned up. We should not explictly drop the reference
on error and also drop the reference when the group is cleaned up. The
new lifetime rules are that an inotify group lives from inotify_new_group
to the last fsnotify_put_group. Since the struct user and inotify_devs
are directly tied to this lifetime they are only changed/updated in
those two locations. We get rid of all special casing of struct
user or user->inotify_devs.
Signed-off-by: Eric Paris <eparis@redhat.com>
Cc: stable@kernel.org (2.6.37 and up)
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

Signed-off-by: yanjun.zhu <yanjun.zhu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>